Mapping functions moved to dedicated utils module
