+++ rewardsweb/core/tests/test_commands.py
@@ -19,6 +19,9 @@ class TestExcel2DbCommand:
         mocked_import = mocker.patch(
             "core.management.commands.excel2db.import_from_csv", return_value=False
         )
+        mocked_map = mocker.patch(
+            "core.management.commands.excel2db.map_github_issues", return_value=False
+        )
         with mock.patch(
             "django.core.management.base.OutputWrapper.write"
         ) as output_log:
@@ -26,31 +29,34 @@ class TestExcel2DbCommand:
             calls = [
                 mocker.call("CSV successfully exported into contributions.csv file!"),
                 mocker.call("Database successfully recreated!"),
+                mocker.call("Issues successfully mapped!"),
             ]
             output_log.assert_has_calls(calls, any_order=True)
-            assert output_log.call_count == 2
+            assert output_log.call_count == 3
 
         fixtures_dir = settings.BASE_DIR.parent / "fixtures"
-        mocked_convert.assert_called_once()
-        mocked_convert.assert_called_with(
+        mocked_convert.assert_called_once_with(
             fixtures_dir / "contributions.xlsx",
             fixtures_dir / "contributions.csv",
             fixtures_dir / "legacy_contributions.csv",
         )
-        mocked_import.assert_called_once()
-        mocked_import.assert_called_with(
+        mocked_import.assert_called_once_with(
             fixtures_dir / "contributions.csv",
             fixtures_dir / "legacy_contributions.csv",
-            github_token="",
         )
+        mocked_map.assert_called_once_with(github_token="")
 
     def test_excel2db_command_output_for_default_values_on_response(self, mocker):
         mocked_convert = mocker.patch(
             "core.management.commands.excel2db.convert_and_clean_excel"
         )
-        response = "response"
+        response1, response2 = "response1", "response2"
         mocked_import = mocker.patch(
-            "core.management.commands.excel2db.import_from_csv", return_value=response
+            "core.management.commands.excel2db.import_from_csv", return_value=response1
+        )
+        mocked_map = mocker.patch(
+            "core.management.commands.excel2db.map_github_issues",
+            return_value=response2,
         )
         with mock.patch(
             "django.core.management.base.OutputWrapper.write"
@@ -58,24 +64,23 @@ class TestExcel2DbCommand:
             call_command("excel2db")
             calls = [
                 mocker.call("CSV successfully exported into contributions.csv file!"),
-                mocker.call(response),
+                mocker.call(response1),
+                mocker.call(response2),
             ]
             output_log.assert_has_calls(calls, any_order=True)
-            assert output_log.call_count == 2
+            assert output_log.call_count == 3
 
         fixtures_dir = settings.BASE_DIR.parent / "fixtures"
-        mocked_convert.assert_called_once()
-        mocked_convert.assert_called_with(
+        mocked_convert.assert_called_once_with(
             fixtures_dir / "contributions.xlsx",
             fixtures_dir / "contributions.csv",
             fixtures_dir / "legacy_contributions.csv",
         )
-        mocked_import.assert_called_once()
-        mocked_import.assert_called_with(
+        mocked_import.assert_called_once_with(
             fixtures_dir / "contributions.csv",
             fixtures_dir / "legacy_contributions.csv",
-            github_token="",
         )
+        mocked_map.assert_called_once_with(github_token="")
 
     def test_excel2db_command_output_for_provided_arguments(self, mocker):
         mocked_convert = mocker.patch(
@@ -83,6 +88,9 @@ class TestExcel2DbCommand:
         )
         mocked_import = mocker.patch(
             "core.management.commands.excel2db.import_from_csv", return_value=False
+        )
+        mocked_map = mocker.patch(
+            "core.management.commands.excel2db.map_github_issues", return_value=False
         )
         input_file, output_file, legacy_file, token = (
             "input_file",
@@ -103,14 +111,12 @@ class TestExcel2DbCommand:
             calls = [
                 mocker.call(f"CSV successfully exported into {output_file} file!"),
                 mocker.call("Database successfully recreated!"),
+                mocker.call("Issues successfully mapped!"),
             ]
             output_log.assert_has_calls(calls, any_order=True)
-            assert output_log.call_count == 2
-        mocked_convert.assert_called_once()
-        mocked_convert.assert_called_with(
+            assert output_log.call_count == 3
+        mocked_convert.assert_called_once_with(
             Path(input_file), Path(output_file), Path(legacy_file)
         )
-        mocked_import.assert_called_once()
-        mocked_import.assert_called_with(
-            Path(output_file), Path(legacy_file), github_token=token
-        )
+        mocked_import.assert_called_once_with(Path(output_file), Path(legacy_file))
+        mocked_map.assert_called_once_with(github_token=token)
